// src/modules/enum/command-tel.enum.ts
export enum ECommandTel {
  HELP = "HELP",
  SEE_MENU = "SEE_MENU",
  GET_ALL_MENU = "GET_ALL_MENU",
  GET_COFFEE_MENU = "GET_COFFEE_MENU",
  GET_CAKE_MENU = "GET_CAKE_MENU",
  GET_TEA_MENU = "GET_TEA_MENU",
  GET_ICE_MENU = "GET_ICE_MENU",
  GET_SMOOTI_MENU = "GET_SMOOTI_MENU",
  GET_ENERGIC_MENU = "GET_ENERGIC_MENU",
  GET_DAMNOOSH_MENU = "GET_DAMNOOSH_MENU",
  GET_SANDWICH_MENU = "GET_SANDWICH_MENU",
  GET_SHACK_MENU = "GET_SHACK_MENU",
  GET_SALAD_MENU = "GET_SALAD_MENU",
  GET_CHOCOLATE_MENU = "GET_CHOCOLATE_MENU",
  GET_SHIRINI_MENU = "GET_SHIRINI_MENU",
  GET_FRAPE_MENU = "GET_FRAPE_MENU",
  GET_NOOSHABE_MENU = "GET_NOOSHABE_MENU",
  GET_BREAKFAST_MENU = "GET_BREAKFAST_MENU",
  GET_BREAD_MENU = "GET_BREAD_MENU",
  GET_COLD_DRINK_MENU = "GET_COLD_DRINK_MENU",
  GET_PASTA_MENU = "GET_PASTA_MENU",
  GET_HOT_DRINK_MENU = "GET_HOT_DRINK_MENU",
  GET_COOKIE_MENU = "GET_COOKIE_MENU",
  GET_MOKTEL_MENU = "GET_MOKTEL_MENU",
}